src/requirements.txt
asgiref==3.2.3
Babel==2.7.0
boto3==1.10.39
botocore==1.13.39
Django==3.0
django-phonenumber-field==4.0.0
djangorestframework==3.11.0
docutils==0.15.2
jmespath==0.9.4
phonenumbers==8.11.1
python-dateutil==2.8.0
pytz==2019.3
s3transfer==0.2.1
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
